third_party/WebKit/LayoutTests/imported/wpt/html/semantics/embedded-content/media-elements/interfaces/TextTrack/kind-expected.txt

This is a testharness.js-based test.
PASS TextTrack.kind, addTextTrack 
PASS TextTrack.kind, track element 
FAIL TextTrack.kind, \u0000 assert_equals: expected "subtitles" but got "metadata"
Harness: the test ran to completion.
